将类型移至相应的包内
