--- packages/wujie-typings/jsx.d.ts
@@ -1,11 +1,0 @@
-import { WujieProps } from './component';
-
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      ['wujie-vue']: WujieProps
-    }
-  }
-}
-
-export {}
