--- packages/wujie-typings/vue26.d.ts
@@ -1,7 +1,0 @@
-declare module '@vue/runtime-dom' {
-  export interface GlobalComponents {
-    WujieVue: typeof import('./component')['WujieComponent']
-  }
-}
-
-export {}
