--- packages/wujie-typings/vue27.d.ts
@@ -1,7 +1,0 @@
-declare module 'vue' {
-  export interface GlobalComponents {
-    WujieVue: typeof import('./component')['WujieComponent']
-  }
-}
-
-export {}
