--- packages/wujie-typings/vue3.d.ts
@@ -1,7 +1,0 @@
-declare module '@vue/runtime-core' {
-  export interface GlobalComponents {
-    WujieVue: typeof import('./component')['WujieComponent']
-  }
-}
-
-export {}
